Home groupings (#2916)
* remove nested

* animations

* blog card

* move hero bottom

* hero bottom

* benefits

* testimonials

* move footer

* nav bar footer

* move button

* fix lint
