+++ src/pages/Home/Animation/StepsCard.tsx
