+++ src/pages/Home/Blogs/BlogCard.tsx
