+++ src/pages/Home/Blogs/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./Blogs";
