+++ src/pages/Home/HeroBottom/HeroBottom.tsx
@@ -1,24 +1,28 @@
 import { appRoutes } from "constants/routes";
-import useCards from "pages/Marketplace/Cards/useCards";
 import { Link } from "react-router-dom";
+import { useEndowmentCardsQuery } from "services/aws/aws";
 import "swiper/css";
 import "swiper/css";
 import "swiper/css/navigation";
 import "swiper/css/pagination";
 import { Navigation } from "swiper/modules";
 import { Swiper, SwiperSlide } from "swiper/react";
-import frame1 from "../../assets/landing/frame_1.png";
-import frame2 from "../../assets/landing/frame_2.png";
-import frame3 from "../../assets/landing/frame_3.png";
-import frame4 from "../../assets/landing/frame_4.png";
-import frame5 from "../../assets/landing/frame_5.png";
-import frame6 from "../../assets/landing/frame_6.png";
-import mappng from "../../assets/landing/map.png";
-import Button from "./Button";
+import frame1 from "../../../assets/landing/frame_1.png";
+import frame2 from "../../../assets/landing/frame_2.png";
+import frame3 from "../../../assets/landing/frame_3.png";
+import frame4 from "../../../assets/landing/frame_4.png";
+import frame5 from "../../../assets/landing/frame_5.png";
+import frame6 from "../../../assets/landing/frame_6.png";
+import mappng from "../../../assets/landing/map.png";
+import Button from "../common/Button";
 import Card from "./Card";
 
 const HeroBottom = () => {
-  const { data } = useCards();
+  const { data } = useEndowmentCardsQuery({
+    claimed: "true",
+    page: 1,
+    query: "",
+  });
 
   return (
     <div className="relative bottom-[300px] lg:bottom-[300px]   md:bottom-[350px] z-10 pt-[50px] px-6 ">
