+++ src/pages/Home/NewNavbar.tsx
@@ -1,7 +1,7 @@
 import { DappLogo } from "components/Image";
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import "../../pages/Home/App.css";
+import "./App.css";
 const NewNavbar = () => {
   const [search, setSearch] = useState("");
   const navigate = useNavigate();
