+++ src/pages/Home/Testimonials/Testimonials.tsx
@@ -2,9 +2,9 @@ import "swiper/css";
 import "swiper/css/pagination";
 import { Navigation } from "swiper/modules";
 import { Swiper, SwiperSlide } from "swiper/react";
-import quotation from "../../../../assets/landing/quotation.svg";
-import TestimonialCard from "../../../../components/landing/TestimonialCard";
-import useGetTestimonials from "../../../../constants/useGetTestimonials";
+import quotation from "../../../assets/landing/quotation.svg";
+import TestimonialCard from "./TestimonialCard";
+import useGetTestimonials from "./useGetTestimonials";
 
 const Testimonials = () => {
   const testimonials = useGetTestimonials();
