+++ src/pages/Home/Testimonials/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./Testimonials";
