+++ src/pages/Home/Testimonials/useGetTestimonials.ts
@@ -1,10 +1,10 @@
-import casd from "../assets/landing/CASD-SL.webp";
-import ph from "../assets/landing/PH8_Logo.jpeg";
-import ace from "../assets/landing/ace_of_hearts.webp";
-import buck from "../assets/landing/buck.webp";
-import fora from "../assets/landing/foraday.png";
-import shoe from "../assets/landing/shoes_that_speak.webp";
-import test from "../assets/landing/testi3Img.png";
+import casd from "assets/landing/CASD-SL.webp";
+import ph from "assets/landing/PH8_Logo.jpeg";
+import ace from "assets/landing/ace_of_hearts.webp";
+import buck from "assets/landing/buck.webp";
+import fora from "assets/landing/foraday.png";
+import shoe from "assets/landing/shoes_that_speak.webp";
+import test from "assets/landing/testi3Img.png";
 export default function useGetTestimonials() {
   return [
     {
